UNIPROTKB {UniProt.ws}	R Documentation

Descriptions of available values for columns and keytypes.

Description

This manual page enumerates the kinds of data represented by the values returned when the user calls columns or keytypes

Details

All the possible values for columns and keytypes are listed below. Users will have to actually use these methods to learn which of the following possible values actually apply in their case.

:UNIPROTKB

The central ID for UniProt and swissprot

:UNIPARC

UniParc

:UNIREF50

UniRef50

:UNIREF90

UniRef90

:UNIREF100

UniRef100

:EMBL/GENBANK/DDBJ

EMBL/GenBank/DDBJ

:EMBL/GENBANK/DDBJ_CDS

EMBL/GenBank/DDBJ CDS

:PIR

PIR

:UNIGENE

UniGene

:ENTREZ_GENE

Entrez Gene (GeneID)

:GI_NUMBER*

GI number

:IPI

IPI

:REFSEQ_PROTEIN

RefSeq Protein

:REFSEQ_NUCLEOTIDE

RefSeq Nucleotide

:PDB

PDB

:DISPROT

DisProt

:HSSP

HSSP

:DIP

DIP

:MINT

MINT

:ALLERGOME

Allergome

:MEROPS

MEROPS

:PEROXIBASE

PeroxiBase

:PPTASEDB

PptaseDB

:REBASE

REBASE

:TCDB

TCDB

:PHOSSITE

PhosSite

:DMDM

DMDM

:AARHUS/GHENT-2DPAGE

Aarhus/Ghent-2DPAGE

:ECO2DBASE

ECO2DBASE

:WORLD-2DPAGE

World-2DPAGE

:DNASU

DNASU

:ENSEMBL

Ensembl

:ENSEMBL_PROTEIN

Ensembl Protein

:ENSEMBL_TRANSCRIPT

Ensembl Transcript

:ENSEMBL_GENOMES

Ensembl Genomes

:ENSEMBL_GENOMES PROTEIN

Ensembl Genomes Protein

:ENSEMBL_GENOMES TRANSCRIPT

Ensembl Genomes Transcript

:KEGG

KEGG

:PATRIC

PATRIC

:TIGR

TIGR

:UCSC

UCSC

:VECTORBASE

VectorBase

:AGD

AGD

:ARACHNOSERVER

ArachnoServer

:CGD

CGD

:CONOSERVER

ConoServer

:CYGD

CYGD

:DICTYBASE

dictyBase

:ECHOBASE

EchoBASE

:ECOGENE

EcoGene

:EUHCVDB

euHCVdb

:EUPATHDB

EuPathDB

:FLYBASE

FlyBase

:GENECARDS

GeneCards

:GENEFARM

GeneFarm

:GENOLIST

GenoList

:H-INVDB

H-InvDB

:HGNC

HGNC

:HPA

HPA

:LEGIOLIST

LegioList

:LEPROMA

Leproma

:MAIZEGDB

MaizeGDB

:MIM

MIM

:MGI

MGI

:NEXTPROT

neXtProt

:ORPHANET

Orphanet

:PHARMGKB

PharmGKB

:POMBASE

PomBase

:PSEUDOCAP

PseudoCAP

:RGD

RGD

:SGD

SGD

:TAIR

TAIR

:TUBERCULIST

TubercuList

:WORMBASE

WormBase

:WORMBASE_TRANSCRIPT

WormBase Transcript

:WORMBASE_PROTEIN

WormBase Protein

:XENBASE

Xenbase

:ZFIN

ZFIN

:EGGNOG

eggNOG

:GENETREE

GeneTree

:HOVERGEN

HOVERGEN

:KO

KO

:OMA

OMA

:ORTHODB

OrthoDB

:PROTCLUSTDB

ProtClustDB

:BIOCYC

BioCyc

:REACTOME

Reactome

:UNIPATHWAY

UniPathWay

:CLEANEX

CleanEx

:GERMONLINE

GermOnline

:DRUGBANK

DrugBank

:GENOMERNAI

GenomeRNAi

:NEXTBIO

NextBio

:CITATION

citations

:CLUSTERS

clusters

:COMMENTS

comments

:DOMAINS

domains

:DOMAIN

domain

:EC

ec ID

:ID

ID

:EXISTENCE

existence

:FAMILIES

families

:FEATURES

features

:GENES

genes

:GO

go term

:GO-ID

go id

:INTERPRO

interpro

:INTERACTOR

interactor

:KEYWORDS

keywords

:KEYWORD-ID

keyword-id

:LAST-MODIFIED

last-modified

:LENGTH

length

:ORGANISM

organism

:ORGANISM-ID

organism-id

:PATHWAY

pathway

:PROTEIN NAMES

protein names

:REVIEWED

reviewed

:SCORE

score

:SEQUENCE

sequence

:3D

3d

:TAXON

taxon

:TOOLS

tools

:VERSION

version

:DATABASE(PFAM)

PFAM ids

:DATABASE(PDB)

PDB ids

:

Author(s)

Marc Carlson

Examples

  up <- UniProt.ws(taxId=9606)
  ## List the possible values for columns
  columns(up)
  ## List the possible values for keytypes
  keytypes(up)
  ## get some values back
  ## list all possible keys of type entrez gene ID.
  ## (this process is not instantaneous)
  if(interactive()){
    keys <- head(keys(up, keytype="UNIPROTKB"))
    keys
  }
  select(up, keys=c("P31946","P62258"), columns=c("PDB","SEQUENCE"),
  keytype="UNIPROTKB")

[Package UniProt.ws version 2.22.0 Index]