utilities {UniProt.ws}	R Documentation

Utility functions

Description

UniProt uses custom coding of organism names from which protein sequences they store. These taxon names are used also in the protein names (not in the UniProt IDs!). These functions help to translate those names to standard scientific (Latin) taxon names and other useful identifiers.

taxname2species(): converts UniProt taxonomy names to scientific species names

taxname2taxid(): converts UniProt taxonomy names to NCBI Taxonomy IDs

taxname2domain(): converts UniProt taxonomy names to the following taxonomical domains: 'A' for archaea (=archaebacteria)\ 'B' for bacteria (=prokaryota or eubacteria)\ 'E' for eukaryota (=eukarya)\ 'V' for viruses and phages (=viridae)\ 'O' for others (such as artificial sequences)\

updatespecfile(): The updatespecfile helper function attempts to download the current version of the controlled vocabulary of species table from UniProt controlled vocabulary of species. If it fails to download, an archived version of the table in (in extdata/) will be used.

Usage

taxname2species(taxname, specfile) 
taxname2taxid(taxname, specfile) 
taxname2domain(taxname, specfile) 
updatespecfile()

Arguments

taxname
Character string up to 6 uppercase characters, like HUMAN, MOUSE, or AERPX. Also works for a vector of such taxon names.

specfile
An optional local file where speclist.RData is saved from UniProt.org. When specfile is missing, a cached file from the extdata/ package directory is used.

Value

Function taxname2species returns a character vector of scientific taxon names matching to the UniProt taxon names supplied as taxname.

Function taxname2taxid returns a numeric vector of Taxonomy IDs matching to the UniProt taxon names supplied as taxname.

Function taxname2domain returns a character vector of one letter domain symbols matching to the UniProt taxon names supplied as taxname.

Function updatespecfile returns a file location where specfile.txt is downloaded. If the download fails, path is the location of the archived version in the package.

Author(s)

Csaba Ortutay

See Also

UniProt controlled vocabulary of species, which defines the taxon names.

Examples

taxname2species("PIG")
taxname2species(c("PIG","HUMAN","TRIHA"))

taxname2taxid("PIG")
taxname2taxid(c("PIG","HUMAN","TRIHA"))

taxname2domain("PIG")
taxname2domain(c("PIG","HUMAN","TRIHA"))

newspecfile <- updatespecfile()
taxname2domain("PIG", specfile = newspecfile)
taxname2domain(c("PIG","HUMAN","TRIHA"), specfile = newspecfile)

[Package UniProt.ws version 2.22.0 Index]